Q key now ctrl, tab alt, tilde caps.  Move many symbols.
Caps lock on a separate level would be better, but unless I fix bugs in
xkb (I might) it won't happen.  But I don't like pressing the tilde key,
because it's just too far (although I don't feel so bad about hitting
backspace... mostly because I move my whole hand for that).  So tilde is
a good candidate to be easily pressable but not in the way or in a
better place than caps should be.  Q is even better than tab for ctrl,
and alt is now in a comfy place.  But this necessitated the moving of
many symbols.  We'll see if I like their new places.  Maybe they'll move
another time or two.